# docs/deliverables/system_docs_20260729/star-rail-agents.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Star Rail Agents",
    "description": "基于 Multi-Agent 与 RAG 的游戏智能助手平台",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/teams/recommendations": {
      "post": {
        "tags": [
          "teams"
        ],
        "summary": "根据核心角色和用户角色池生成官方角色配队",
        "operationId": "recommend_official_teams_api_v1_teams_recommendations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfficialTeamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfficialTeamResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/profile/teams": {
      "get": {
        "tags": [
          "teams"
        ],
        "summary": "List Saved Teams",
        "operationId": "list_saved_teams_api_v1_profile_teams_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedTeamListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "teams"
        ],
        "summary": "Save Team",
        "operationId": "save_team_api_v1_profile_teams_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedTeamCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedTeamResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/profile/teams/{team_id}": {
      "patch": {
        "tags": [
          "teams"
        ],
        "summary": "Rename Saved Team",
        "operationId": "rename_saved_team_api_v1_profile_teams__team_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Team Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedTeamUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedTeamResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "teams"
        ],
        "summary": "Delete Saved Team",
        "operationId": "delete_saved_team_api_v1_profile_teams__team_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIResponse": {
        "properties": {
          "response_id": {
            "type": "string",
            "format": "uuid",
            "title": "Response Id"
          },
          "agent": {
            "type": "string",
            "title": "Agent"
          },
          "protocol": {
            "type": "string",
            "const": "claim-citation-validation-filtering",
            "title": "Protocol",
            "default": "claim-citation-validation-filtering"
          },
          "answer": {
            "type": "string",
            "title": "Answer",
            "default": ""
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/Claim"
            },
            "type": "array",
            "title": "Claims"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations"
          },
          "validation": {
            "$ref": "#/components/schemas/ValidationReport"
          },
          "filtering": {
            "$ref": "#/components/schemas/FilteringReport"
          },
          "query_steps": {
            "items": {
              "$ref": "#/components/schemas/QueryStep"
            },
            "type": "array",
            "title": "Query Steps"
          },
          "invoked_agents": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Invoked Agents"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "memory_suggestions": {
            "items": {
              "$ref": "#/components/schemas/MemorySuggestion"
            },
            "type": "array",
            "title": "Memory Suggestions"
          },
          "favorite_character_suggestions": {
            "items": {
              "$ref": "#/components/schemas/FavoriteCharacterSuggestion"
            },
            "type": "array",
            "title": "Favorite Character Suggestions"
          },
          "related_entities": {
            "items": {
              "$ref": "#/components/schemas/app__schemas__ai_response__RelatedEntity"
            },
            "type": "array",
            "title": "Related Entities"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/AgentAction"
            },
            "type": "array",
            "title": "Actions"
          },
          "follow_up_questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 3,
            "title": "Follow Up Questions"
          }
        },
        "type": "object",
        "required": [
          "agent",
          "claims",
          "citations",
          "validation",
          "filtering"
        ],
        "title": "AIResponse"
      },
...
      "Citation": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "excerpt": {
            "type": "string",
            "title": "Excerpt",
            "default": ""
          },
          "document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "chunk_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Id"
          },
          "game_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Game Version"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "relevance_score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Relevance Score"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "entity_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "source"
        ],
        "title": "Citation"
      },
      "Claim": {
        "properties": {
          "statement": {
            "type": "string",
            "title": "Statement"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "citation_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Citation Ids"
          },
          "claim_type": {
            "type": "string",
            "enum": [
              "fact",
              "interpretation"
            ],
            "title": "Claim Type",
            "default": "fact"
          },
          "evidence_quotes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Quotes"
          }
        },
        "type": "object",
        "required": [
          "statement",
          "confidence"
        ],
        "title": "Claim"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OfficialRecommendedTeam": {
        "properties": {
          "members": {
            "items": {
              "$ref": "#/components/schemas/OfficialTeamMember"
            },
            "type": "array",
            "title": "Members"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "covered_roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Covered Roles"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Strengths"
          },
          "weaknesses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Weaknesses"
          },
          "source_character_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Character Ids"
          },
          "preferred_character_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Preferred Character Ids"
          },
          "missing_character_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Character Ids"
          },
          "model_assessment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Assessment"
          },
          "model_verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Verdict"
          },
          "score_breakdown": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TeamScoreBreakdown"
              },
              {
                "type": "null"
              }
            ]
          },
          "rotation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TeamRotationSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "data_warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Data Warnings"
          }
        },
        "type": "object",
        "required": [
          "members",
          "score",
          "covered_roles",
          "reasons",
          "strengths",
          "weaknesses",
          "source_character_ids"
        ],
        "title": "OfficialRecommendedTeam"
      },
      "OfficialTeamMember": {
        "properties": {
          "character_id": {
            "type": "string",
            "title": "Character Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "element": {
            "type": "string",
            "title": "Element"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles"
          },
          "mechanic_tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Mechanic Tags"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          }
        },
        "type": "object",
        "required": [
          "character_id",
          "name",
          "element",
          "roles"
        ],
        "title": "OfficialTeamMember"
      },
      "OfficialTeamRequest": {
        "properties": {
          "core_character_id": {
            "type": "string",
            "maxLength": 16,
            "minLength": 1,
            "title": "Core Character Id"
          },
          "preferred_character_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Preferred Character Ids"
          },
          "excluded_character_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Excluded Character Ids"
          },
          "require_sustain": {
            "type": "boolean",
            "title": "Require Sustain",
            "default": true
          },
          "use_owned_only": {
            "type": "boolean",
            "title": "Use Owned Only",
            "default": false
          },
          "game_mode": {
            "type": "string",
            "enum": [
              "balanced",
              "moc",
              "pure_fiction",
              "apocalyptic"
            ],
            "title": "Game Mode",
            "default": "balanced"
          }
        },
        "type": "object",
        "required": [
          "core_character_id"
        ],
        "title": "OfficialTeamRequest"
      },
      "OfficialTeamResponse": {
        "properties": {
          "response": {
            "$ref": "#/components/schemas/AIResponse"
          },
          "theoretical": {
            "items": {
              "$ref": "#/components/schemas/OfficialRecommendedTeam"
            },
            "type": "array",
            "title": "Theoretical"
          },
          "owned": {
            "items": {
              "$ref": "#/components/schemas/OfficialRecommendedTeam"
            },
            "type": "array",
            "title": "Owned"
          },
          "favorite_trials": {
            "items": {
              "$ref": "#/components/schemas/OfficialRecommendedTeam"
            },
            "type": "array",
            "title": "Favorite Trials"
          }
        },
        "type": "object",
        "required": [
          "response",
          "theoretical",
          "owned"
        ],
        "title": "OfficialTeamResponse"
      },
...
      "SavedTeamCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 48,
            "minLength": 1,
            "title": "Name"
          },
          "member_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 4,
            "title": "Member Ids"
          }
        },
        "type": "object",
        "required": [
          "name",
          "member_ids"
        ],
        "title": "SavedTeamCreate"
      },
      "SavedTeamListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SavedTeamResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "SavedTeamListResponse"
      },
      "SavedTeamResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/OfficialTeamMember"
            },
            "type": "array",
            "title": "Members"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "members",
          "created_at",
          "updated_at"
        ],
        "title": "SavedTeamResponse"
      },
      "SavedTeamUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 48,
            "minLength": 1,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "SavedTeamUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "本地 FastAPI"
    }
  ]
}